move janki model to dedicated file
